Add session_manager.py and move some session-related classes into it
