+++ lib/streamlit/runtime/runtime.py
@@ -40,6 +40,7 @@ from .legacy_caching.caching import _mem_caches
 from .media_file_manager import MediaFileManager
 from .media_file_storage import MediaFileStorage
 from .session_data import SessionData
+from .session_manager import SessionClient, SessionClientDisconnectedError, SessionInfo
 from .state import SessionStateStatProvider, SCRIPT_RUN_WITHOUT_ERRORS_KEY
 from .stats import StatsManager
 from .uploaded_file_manager import UploadedFileManager
@@ -50,23 +51,8 @@ SCRIPT_RUN_CHECK_TIMEOUT: Final = 60
 LOGGER: Final = get_logger(__name__)
 
 
-class SessionClientDisconnectedError(Exception):
-    """Raised by operations on a disconnected SessionClient."""
-
-
 class RuntimeStoppedError(Exception):
     """Raised by operations on a Runtime instance that is stopped."""
-
-
-class SessionClient(Protocol):
-    """Interface for sending data to a session's client."""
-
-    def write_forward_msg(self, msg: ForwardMsg) -> None:
-        """Deliver a ForwardMsg to the client.
-
-        If the SessionClient has been disconnected, it should raise a
-        SessionClientDisconnectedError.
-        """
 
 
 class RuntimeConfig(NamedTuple):
@@ -81,29 +67,6 @@ class RuntimeConfig(NamedTuple):
 
     # The storage backend for Streamlit's MediaFileManager.
     media_file_storage: MediaFileStorage
-
-
-class SessionInfo:
-    """Type stored in our _session_info_by_id dict.
-
-    For each AppSession, the Runtime tracks that session's
-    script_run_count. This is used to track the age of messages in
-    the ForwardMsgCache.
-    """
-
-    def __init__(self, client: SessionClient, session: AppSession):
-        """Initialize a SessionInfo instance.
-
-        Parameters
-        ----------
-        session : AppSession
-            The AppSession object.
-        client : SessionClient
-            The concrete SessionClient for this session.
-        """
-        self.session = session
-        self.client = client
-        self.script_run_count = 0
 
 
 class RuntimeState(Enum):
